Cleaned up utils, merged, and blackened
