+++ vectorizers/utils.py
@@ -4,15 +4,12 @@ import scipy.stats
 import scipy.sparse
 import itertools
 from collections.abc import Iterable
-from collections import namedtuple
 from sklearn.mixture import GaussianMixture
 from sklearn.preprocessing import normalize, LabelBinarizer
 from collections import Counter
 import re
 from warnings import warn
 from numba.np.unsafe.ndarray import to_fixed_tuple
-
-CooArray = namedtuple("CooArray", ["row", "col", "val", "key", "ind", "min"])
 
 
 @numba.njit(nogil=True)
@@ -26,6 +23,7 @@ def make_tuple_converter(tuple_size):
         return to_fixed_tuple(to_convert, tuple_size)
 
     return tuple_converter
+
 
 def str_to_bytes(size_str):
     """Convert a string description of a memory size (e.g. 1GB, or 100M, or 20kB)
@@ -71,113 +69,12 @@ def str_to_bytes(size_str):
         return int(np.ceil(float(parse_match.group(1))))
 
 
-@numba.njit(nogil=True, inline="always")
-def coo_sum_duplicates(coo, kind):
-    upper_lim = coo.ind[0]
-    lower_lim = coo.min[0]
-
-    perm = np.argsort(coo.key[lower_lim:upper_lim], kind=kind)
-
-    coo.row[lower_lim:upper_lim] = coo.row[lower_lim:upper_lim][perm]
-    coo.col[lower_lim:upper_lim] = coo.col[lower_lim:upper_lim][perm]
-    coo.val[lower_lim:upper_lim] = coo.val[lower_lim:upper_lim][perm]
-    coo.key[lower_lim:upper_lim] = coo.key[lower_lim:upper_lim][perm]
-
-    sum_ind = lower_lim
-    this_row = coo.row[lower_lim]
-    this_col = coo.col[lower_lim]
-    this_val = np.float32(0)
-    this_key = coo.key[lower_lim]
-
-    for i in range(lower_lim, upper_lim):
-        if coo.key[i] == this_key:
-            this_val += coo.val[i]
-        else:
-            coo.row[sum_ind] = this_row
-            coo.col[sum_ind] = this_col
-            coo.val[sum_ind] = this_val
-            coo.key[sum_ind] = this_key
-            this_row = coo.row[i]
-            this_col = coo.col[i]
-            this_val = coo.val[i]
-            this_key = coo.key[i]
-            sum_ind += 1
-
-    if this_key != coo.key[upper_lim]:
-        coo.row[sum_ind] = this_row
-        coo.col[sum_ind] = this_col
-        coo.val[sum_ind] = this_val
-        coo.key[sum_ind] = this_key
-        sum_ind += 1
-
-    coo.ind[0] = sum_ind
-    coo.min[0] = coo.ind[0]
-
-
-@numba.njit(nogil=True, inline="always")
-def coo_append(coo, tup):
-    coo.row[coo.ind[0]] = tup[0]
-    coo.col[coo.ind[0]] = tup[1]
-    coo.val[coo.ind[0]] = tup[2]
-    coo.key[coo.ind[0]] = tup[3]
-    coo.ind[0] += 1
-
-    if (coo.ind[0] - coo.min[0]) >= 1 << 18:
-        coo_sum_duplicates(coo, kind="quicksort")
-        if coo.key.shape[0] - coo.min[0] <= 1 << 18:
-            coo.min[0] = 0.0
-            coo_sum_duplicates(coo, kind="mergesort")
-            if coo.ind[0] >= 0.95 * coo.key.shape[0]:
-                raise ValueError(
-                    f"The coo matrix array is over memory limit.  Increase coo_max_bytes to process data."
-                )
-
-    if coo.ind[0] == coo.key.shape[0]:
-        coo.min[0] = 0.0
-        coo_sum_duplicates(coo, kind="mergesort")
-        if coo.ind[0] >= 0.95 * coo.key.shape[0]:
-            raise ValueError(
-                f"The coo matrix array is over memory limit.  Increase coo_max_bytes to process data."
-            )
-
-
 def flatten(list_of_seq):
     assert isinstance(list_of_seq, Iterable)
     if type(list_of_seq[0]) in (list, tuple, np.ndarray):
         return tuple(itertools.chain.from_iterable(list_of_seq))
     else:
         return list_of_seq
-
-
-@numba.njit(nogil=True)
-def sum_coo_entries(seq):
-    seq.sort()
-    this_coord = (seq[0][0], seq[0][1])
-    this_sum = 0
-    reduced_data = []
-    for entry in seq:
-        if (entry[0], entry[1]) == this_coord:
-            this_sum += entry[2]
-        else:
-            reduced_data.append((this_coord[0], this_coord[1], this_sum))
-            this_sum = entry[2]
-            this_coord = (entry[0], entry[1])
-
-    reduced_data.append((this_coord[0], this_coord[1], this_sum))
-
-    return reduced_data
-
-
-@numba.njit(nogil=True)
-def update_coo_entries(seq, tup):
-    place = np.searchsorted(seq, tup)
-    if seq[place][1:2] == tup[1:2]:
-        seq[place][3] += tup[3]
-        return seq
-    elif seq[place - 1][1:2] == tup[1:2]:
-        seq[place - 1][3] += tup[3]
-        return seq
-    return seq.insert(place, tup)
 
 
 def sparse_collapse(matrix, labels, sparse=True):
@@ -297,7 +194,7 @@ def gmm_component_likelihood(
 
 
 def vectorize_diagram(diagram: np.ndarray, gmm: GaussianMixture) -> np.ndarray:
-    """Given a diagram and a Guassian Mixture Model, produce the vectorized
+    """Given a diagram and a Gaussian Mixture Model, produce the vectorized
     representation of the diagram as a vector of weights associated to
     each component of the GMM.
